Initial move shared outputs to a directory
