+++ src/components/outputs/shared-with-iframe/ImageOutput.tsx
