+++ src/components/outputs/shared-with-iframe/PlainTextOutput.tsx
@@ -9,7 +9,8 @@ export const PlainTextOutput: React.FC<PlainTextOutputProps> = ({
   content,
   className = "font-mono text-sm whitespace-pre-wrap leading-relaxed text-gray-700",
 }) => {
-  return <div className={className}>{content}</div>;
+  // TODO: added stuff for testing
+  return <div className={className}>plain text output {content}</div>;
 };
 
 export default PlainTextOutput;
