+++ src/util/prefetch.ts
@@ -75,14 +75,26 @@ export function prefetchOutputChunks(): void {
 
     // These imports will trigger chunk loading but won't execute the modules
     // until they're actually needed via React.lazy()
-    import("../components/outputs/MarkdownRenderer.js").catch(() => {
+    import(
+      "../components/outputs/shared-with-iframe/MarkdownRenderer.js"
+    ).catch(() => {
       // Silently ignore prefetch failures
     });
-    import("../components/outputs/JsonOutput.js").catch(() => {});
-    import("../components/outputs/PlainTextOutput.js").catch(() => {});
-    import("../components/outputs/HtmlOutput.js").catch(() => {});
-    import("../components/outputs/ImageOutput.js").catch(() => {});
-    import("../components/outputs/SvgOutput.js").catch(() => {});
+    import("../components/outputs/shared-with-iframe/JsonOutput.js").catch(
+      () => {}
+    );
+    import("../components/outputs/shared-with-iframe/PlainTextOutput.js").catch(
+      () => {}
+    );
+    import("../components/outputs/shared-with-iframe/HtmlOutput.js").catch(
+      () => {}
+    );
+    import("../components/outputs/shared-with-iframe/ImageOutput.js").catch(
+      () => {}
+    );
+    import("../components/outputs/shared-with-iframe/SvgOutput.js").catch(
+      () => {}
+    );
     import("../components/outputs/AiToolCallOutput.js").catch(() => {});
     import("../components/outputs/AiToolResultOutput.js").catch(() => {});
   });
@@ -121,8 +133,12 @@ export function prefetchOutputsConservative(): void {
   prefetchWhenIdle(
     () => {
       // Only prefetch the most commonly used components
-      import("../components/outputs/PlainTextOutput.js").catch(() => {});
-      import("../components/outputs/MarkdownRenderer.js").catch(() => {});
+      import(
+        "../components/outputs/shared-with-iframe/PlainTextOutput.js"
+      ).catch(() => {});
+      import(
+        "../components/outputs/shared-with-iframe/MarkdownRenderer.js"
+      ).catch(() => {});
       // Preload react-spring as it's used in the loading screen
       import("@react-spring/web").catch(() => {});
     },
